module5-task1 Update classes with Inheritance
