+++ src/task.js
@@ -1,7 +1,10 @@
-import {tags, createElement} from './utils';
+import {tags} from './utils';
+import Component from './component';
 
-class Task {
+export default class Task extends Component {
+
   constructor(data) {
+    super();
     this._title = data.title;
     this._dueDate = data.dueDate;
     this._tags = data.tags;
@@ -9,17 +12,12 @@ class Task {
     this._color = data.color;
     this._isDone = data.isDone;
     this._isFavorite = data.isFavorite;
-    this._element = null;
     this._onEdit = null;
     this._onEditButtonClick = this._onEditButtonClick.bind(this);
   }
 
   _onEditButtonClick() {
     return typeof this._onEdit === `function` && this._onEdit();
-  }
-
-  get element() {
-    return this._element;
   }
 
   set onEdit(fn) {
@@ -127,27 +125,14 @@ value="${new Date(this._dueDate).toLocaleString(`en-US`, {hour: `2-digit`, minut
     `.trim();
   }
 
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
 
-  unrender() {
-    this.unbind();
-    this._element = null;
-  }
-
-
-  bind() {
+  createListeners() {
     this._element.querySelector(`.card__btn--edit`)
       .addEventListener(`click`, this._onEditButtonClick);
   }
 
-  unbind() {
+  removeListeners() {
     this._element.querySelector(`.card__btn--edit`)
       .removeEventListener(`click`, this._onEditButtonClick);
   }
 }
-
-export default Task;
